pyproject.toml
[tool.poetry]
name = "elektra-back"
version = "0.1.0"
description = ""
authors = ["RobertoAR <roberto@apprecio.mx>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
django = "3.1.7"
djongo = "^1.3.6"
django-cors-headers = "3.7.0"
pymongo = "3.12.3"
dj-rest-auth = "2.1.11"
djangorestframework-simplejwt = "5.0.0"
django-rest-auth = "^0.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
